requirements.txt
Django==3.2.7
djangorestframework==3.12.4
Pillow==8.4.0
pdf2image==1.16.0
asgiref==3.7.2
Django==3.2.7
djangorestframework==3.12.4
pdf2image==1.16.0
Pillow==8.4.0
PyMuPDF==1.22.5
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
